Skill Tree and Save
Skills must be unlocked and set to active. Only 3 skills can be active at a time. When a skill is unlocked, set to active, or deactivated, the player skill tree state and list of active skills is saved, then loaded in when the player next spawns. Setting a skill to active does not immediately apply the skill.